gnuChess for windows, RRDtool, NIC SOA requirement
I looked for a version of gnuChess for windows for a while.
The reason that it took me so long was that I was used to
playing gnugo, which is best with some sort of front-end.
gnuChess comes with one build it!

Munin will graph day,week,month,year; What if you need decade?
What if you need to extract data from a rrd and take an
AVERAGE over a time period.

When registering domain names with some regional NICs,
they can stipulate valid SOA values in your DNS.
These are derived from RFCs and though they can cause
more work for you, there are reasons behind them.
(Which can be useful when a customer wants a 5 min TTL.)