palm application server written in perl
http://monkey-project.com/ seems like the nginx killer.
If you want to server more than just static files you need
a daemon that will run the CGI scripts for you and return the
output. I threw this together tonight in perl because the
one provided with the server, (in python) did not work for me.

I'm off to fix the auth plugin so that it reads .htaccess files
and then I'm going to uninstall apache.